Fix cross-platform ww_download

+++ src/libraries/ww_vegas/ww_download/registry.cpp
@@ -31,8 +31,8 @@ bool getStringFromRegistry(HKEY root, std::string path, std::string key, std::st
 {
 	HKEY handle;
 	unsigned char buffer[256];
-	unsigned long size = 256;
-	unsigned long type;
+        DWORD size = 256;
+        DWORD type;
 	int returnValue;
 
 	if ((returnValue = RegOpenKeyEx(root, path.c_str(), 0, KEY_READ, &handle)) == ERROR_SUCCESS)
@@ -53,9 +53,9 @@ bool getStringFromRegistry(HKEY root, std::string path, std::string key, std::st
 bool getUnsignedIntFromRegistry(HKEY root, std::string path, std::string key, unsigned int &val)
 {
 	HKEY handle;
-	unsigned long buffer;
-	unsigned long size = sizeof(buffer);
-	unsigned long type;
+        unsigned long buffer;
+        DWORD size = sizeof(buffer);
+        DWORD type;
 	int returnValue;
 
 	if ((returnValue = RegOpenKeyEx(root, path.c_str(), 0, KEY_READ, &handle)) == ERROR_SUCCESS)
